requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
apipkg==1.5
astor==0.8.1
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
et-xmlfile==1.1.0
execnet==1.8.0
h11==0.13.0
idna==2.10
iniconfig==1.1.1
openpyxl==3.0.10
outcome==1.2.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.21
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.3
pytest-ast-transformer==1.0.3
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-soft-assertions==0.1.2
pytest-xdist==2.2.1
python-dotenv==0.20.0
requests==2.25.1
selenium==3.141.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
texttable==1.6.3
toml==0.10.2
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.4
webdriver-manager==3.8.3
wsproto==1.2.0
